apps/agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
copilotkit
ag-ui-langgraph
langchain-anthropic
langchain-openai
langchain-core
langgraph
upstash-redis
upstash-vector
pydantic
